e2e_test_platform/requirements.txt
PyYAML>=6.0
colorama>=0.4.6
